+++ src/mantle/console/class-manually-failed-exception.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * Manually_Failed_Exception class file
+ *
+ * @package Mantle
+ */
+
+namespace Mantle\Console;
+
+use RuntimeException;
+
+/**
+ * Manually Failed Command Exception
+ */
+class Manually_Failed_Exception extends RuntimeException {}