+++ CHANGELOG.md
@@ -12,6 +12,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a `Option` and `Object_Metadata` support classes as well as `option()`,
   `post_meta()`, `term_meta()`, `user_meta()`, and `comment_meta()` helpers to
   retrieve options and object metadata in a type-safe manner.
+- Added a `Option` support class and `option()` helper to retrieve options in a type-safe manner.
+- Added `fail()` method to commands to allow for a command to fail with a message.
+
+### Fixed
+
+- Fixed command testing to proper handle failed commands.
 
 ## v1.5.4 - 2025-03-06